+++ Multiplayer-Roguelike.Unity/Assets/Runtime/Ecs/Components/IComponentStorage.cs
@@ -2,12 +2,12 @@ namespace Runtime.Ecs.Components
 {
     public interface IComponentStorage<out T> where T : IComponent
     {
-        int Count { get; }
-        int[] EntityIds { get; }
+        ushort Count { get; }
+        ushort[] EntityIds { get; }
         T[]  Components { get; }
-        bool Has(int entityId);
-        bool TryGet(int id, out IComponent component);
-        T Get(int entityId);
-        void Remove(int entityId);
+        bool Has(ushort entityId);
+        bool TryGet(ushort id, out IComponent component);
+        T Get(ushort entityId);
+        void Remove(ushort entityId);
     }
 }
